Fixed bugs & Targeting Improvements.
Fixed a bug that would cause a problem when attempting to observe attributes/resources in the RPG component. Since only one of the same callback can be present, it creates a freeze temporarily. Now it checks if it exists and only adds it if not.

Added some improvements to the targeting system to hold a cached vector to where it's endpoint of tracing will be.